Split a base class out of the WASM collecting hasher.
This is for readability more than anything, though it'll also be
useful if we ever decide to do multiple hash algorithms.
